backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy==2.0.35
alembic==1.13.1

# Validation & Serialization
pydantic[email]==2.9.2
pydantic-settings==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Environment Variables
python-dotenv==1.0.0

# AI & ML APIs
google-generativeai==0.3.2
openai==1.12.0
anthropic==0.18.1

# Image Processing
Pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.4

# Audio Processing
librosa==0.10.2
soundfile==0.12.1

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# Utilities
python-dateutil==2.8.2
pytz==2024.1
